+++ source/tailscale-utils/usr/local/emhttp/plugins/tailscale/check-tailscale-key.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+/usr/local/emhttp/plugins/tailscale/check-expiration.php | logger -t unraid-tailscale